# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Mini ERP API
  version: 0.1.0
  description: Contract-first MVP skeleton for Mini ERP.
servers:
  - url: /api/v1
...
paths:
...
  /branches/{branch_id}:
    get:
      tags: [Branches]
      summary: Get branch
      operationId: getBranch
      parameters:
        - $ref: '#/components/parameters/BranchId'
      responses:
        '200':
          description: Branch
    patch:
      tags: [Branches]
      summary: Update branch
      operationId: updateBranch
      parameters:
        - $ref: '#/components/parameters/BranchId'
      responses:
        '200':
          description: Updated
...
  /workflows/{workflow_id}:
    get:
      tags: [Workflows]
      summary: Get workflow
      operationId: getWorkflow
      parameters:
        - name: workflow_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Workflow
    patch:
      tags: [Workflows]
      summary: Update workflow
      operationId: updateWorkflow
      parameters:
        - name: workflow_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Workflow updated
  /workflows/{workflow_id}/statuses:
    post:
      tags: [Workflows]
      summary: Add workflow status
      operationId: createWorkflowStatus
      parameters:
        - name: workflow_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: Workflow status created
  /workflows/{workflow_id}/transitions:
    post:
      tags: [Workflows]
      summary: Add workflow transition
      operationId: createWorkflowTransition
      parameters:
        - name: workflow_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: Workflow transition created
...
  /service-orders/{order_id}:
    get:
      tags: [Service Orders]
      summary: Get service order
      operationId: getServiceOrder
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Service order
...
  /resources:
    get:
      tags: [Resources]
      summary: List resources
      operationId: listResources
      parameters:
        - name: branch_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Resource list
    post:
      tags: [Resources]
      summary: Create resource
      operationId: createResource
      responses:
        '201':
          description: Resource created
  /resources/{resource_id}/stock-movements:
    post:
      tags: [Resources]
      summary: Record stock movement
      operationId: recordStockMovement
      parameters:
        - name: resource_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: Stock movement recorded
  /resources/{resource_id}/availability:
    get:
      tags: [Resources]
      summary: Get resource availability
      operationId: getResourceAvailability
      parameters:
        - name: resource_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Resource availability
  /service-orders/{order_id}/resource-usage:
    get:
      tags: [Resources]
      summary: List resource usage for service order
      operationId: listResourceUsage
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Resource usage list
    post:
      tags: [Resources]
      summary: Record resource usage for service order
      operationId: recordResourceUsage
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: Resource usage recorded
...
components:
  parameters:
...
    BranchId:
      name: branch_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...